+++ tests/testthat/test-simulate-secondary.R
@@ -21,7 +21,7 @@ test_that("simulate_secondary works as expected with standard parameters", {
 test_that("simulate_secondary works as expected with additional parameters", {
   set.seed(123)
   sim <- test_simulate_secondary(
-    delays = delay_opts(fix_dist(example_reporting_delay)),
+    delays = delay_opts(fix_parameters(example_reporting_delay)),
     obs = obs_opts(family = "negbin", phi = list(mean = 0.5, sd = 0))
   )
   expect_equal(nrow(sim), nrow(cases))